requirements.txt
django==3.2
djangorestframework==3.12.4
django-url-filter==0.3.15
drf-yasg==1.20.0